test: moved shred test function into test_utils.rs
