+++ src/fhir.rs
@@ -3,4 +3,4 @@ pub(crate) mod location;
 pub(crate) mod mapper;
 pub(crate) mod patient;
 pub(crate) mod resources;
-pub mod sharded_fhir_functions;
+pub mod test_utils;
