+++ src/fhir/patient.rs
@@ -643,7 +643,7 @@ fn field_extension(url: String, ext_value: ExtensionValue) -> Result<FieldExtens
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::fhir::sharded_fhir_functions::get_test_config;
+    use crate::fhir::test_utils::tests::get_test_config;
     use fhir_model::Date;
     use fhir_model::DateTime;
     use fhir_model::r4b::codes::HTTPVerb::Delete;
